+++ include/simdjson/parsedjson.h
@@ -1,3 +1,5 @@
+// TODO Remove this -- deprecated API and files
+
 #ifndef SIMDJSON_PARSEDJSON_H
 #define SIMDJSON_PARSEDJSON_H
