+++ include/simdjson/parsedjsoniterator.h
@@ -1,3 +1,5 @@
+// TODO Remove this -- deprecated API and files
+
 #ifndef SIMDJSON_PARSEDJSONITERATOR_H
 #define SIMDJSON_PARSEDJSONITERATOR_H
