--- include/simdjson/stage1_find_marks.h
@@ -1,6 +1,0 @@
-#ifndef SIMDJSON_STAGE1_FIND_MARKS_H
-#define SIMDJSON_STAGE1_FIND_MARKS_H
-
-#include "simdjson/document.h"
-
-#endif
