--- include/simdjson/stage2_build_tape.h
@@ -1,6 +1,0 @@
-#ifndef SIMDJSON_STAGE2_BUILD_TAPE_H
-#define SIMDJSON_STAGE2_BUILD_TAPE_H
-
-#include "simdjson/document.h"
-
-#endif
