+++ singleheader/simdjson.cpp
@@ -702,8 +702,7 @@ namespace simdjson {
 // instruction sets.
 
 // function pointer type for json_parse
-using json_parse_functype = int(const uint8_t *buf, size_t len, ParsedJson &pj,
-                                bool realloc);
+using json_parse_functype = int(const uint8_t *buf, size_t len, ParsedJson &pj, bool realloc);
 
 // Pointer that holds the json_parse implementation corresponding to the
 // available SIMD instruction set
