+++ src/arm64/stage2_build_tape.h
@@ -5,7 +5,7 @@
 
 #ifdef IS_ARM64
 
-#include "simdjson/stage2_build_tape.h"
+#include "arm64/implementation.h"
 #include "arm64/stringparsing.h"
 #include "arm64/numberparsing.h"
 
@@ -15,22 +15,6 @@ namespace simdjson::arm64 {
 #include "generic/stage2_streaming_build_tape.h"
 
 } // namespace simdjson::arm64
-
-namespace simdjson {
-
-template <>
-WARN_UNUSED int
-unified_machine<architecture::ARM64>(const uint8_t *buf, size_t len, ParsedJson &pj) {
-  return arm64::stage2::unified_machine(buf, len, pj);
-}
-
-template <>
-WARN_UNUSED int
-unified_machine<architecture::ARM64>(const uint8_t *buf, size_t len, ParsedJson &pj, size_t &next_json) {
-    return arm64::stage2::unified_machine(buf, len, pj, next_json);
-}
-
-} // namespace simdjson
 
 #endif // IS_ARM64
