+++ src/haswell/stage2_build_tape.h
@@ -5,7 +5,7 @@
 
 #ifdef IS_X86_64
 
-#include "simdjson/stage2_build_tape.h"
+#include "haswell/implementation.h"
 #include "haswell/stringparsing.h"
 #include "haswell/numberparsing.h"
 
@@ -14,24 +14,6 @@ namespace simdjson::haswell {
 
 #include "generic/stage2_build_tape.h"
 #include "generic/stage2_streaming_build_tape.h"
-
-} // namespace simdjson::haswell
-UNTARGET_REGION
-
-TARGET_HASWELL
-namespace simdjson {
-
-template <>
-WARN_UNUSED int
-unified_machine<architecture::HASWELL>(const uint8_t *buf, size_t len, ParsedJson &pj) {
-  return haswell::stage2::unified_machine(buf, len, pj);
-}
-
-template <>
-WARN_UNUSED int
-unified_machine<architecture::HASWELL>(const uint8_t *buf, size_t len, ParsedJson &pj, UNUSED size_t &next_json) {
-  return haswell::stage2::unified_machine(buf, len, pj, next_json);
-}
 
 } // namespace simdjson
 UNTARGET_REGION
