+++ src/westmere/stage2_build_tape.h
@@ -5,7 +5,7 @@
 
 #ifdef IS_X86_64
 
-#include "simdjson/stage2_build_tape.h"
+#include "westmere/implementation.h"
 #include "westmere/stringparsing.h"
 #include "westmere/numberparsing.h"
 
@@ -16,25 +16,6 @@ namespace simdjson::westmere {
 #include "generic/stage2_streaming_build_tape.h"
 
 } // namespace simdjson::westmere
-UNTARGET_REGION
-
-TARGET_WESTMERE
-namespace simdjson {
-
-template <>
-WARN_UNUSED int
-unified_machine<architecture::WESTMERE>(const uint8_t *buf, size_t len, ParsedJson &pj) {
-  return westmere::stage2::unified_machine(buf, len, pj);
-}
-
-template <>
-WARN_UNUSED int
-unified_machine<architecture::WESTMERE>(const uint8_t *buf, size_t len, ParsedJson &pj, size_t &next_json) {
-    return westmere::stage2::unified_machine(buf, len, pj, next_json);
-}
-
-
-} // namespace simdjson
 UNTARGET_REGION
 
 #endif // IS_X86_64
